--- lib/utils/sizes.js
@@ -1,13 +1,0 @@
-const WIDTH_OPTIONS = {
-  xs: "max-w-xs",
-  sm: "max-w-sm",
-  md: "max-w-md",
-  lg: "max-w-lg",
-  xl: "max-w-xl",
-  "2xl": "max-w-2xl",
-  "3xl": "max-w-3xl",
-  "4xl": "max-w-4xl",
-  "5xl": "max-w-5xl",
-};
-
-export { WIDTH_OPTIONS };
